Adjacency list format has a header line
Input adjacency list now contains a header. It is separated from the body
by a blank line. It consists of a single line with three numbers which are
separated by spaces.

  vertex_count edge_count max_ordinal

Metadata can be used by io.jl to simplify the reader. It is expected that
metadata will be created by a hadoop job.